requirements.txt
PyYAML==6.0.2
requests==2.32.4
loguru==0.7.2